requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
rich>=13.7.1
tenacity>=8.2.3

playwright>=1.45.0

sqlalchemy>=2.0.30
psycopg[binary]>=3.1.19

diffusers>=0.30.0
transformers>=4.43.0
accelerate>=0.33.0
safetensors>=0.4.3
torch>=2.3.0

numpy>=1.26.4
pillow>=10.3.0

